refactor: move modem properties getters
