+++ kura/org.eclipse.kura.nm/src/main/java/org/eclipse/kura/nm/NMDbusConnector.java
@@ -50,7 +50,6 @@ import org.freedesktop.dbus.connections.impl.DBusConnection;
 import org.freedesktop.dbus.exceptions.DBusException;
 import org.freedesktop.dbus.exceptions.DBusExecutionException;
 import org.freedesktop.dbus.interfaces.Properties;
-import org.freedesktop.dbus.types.UInt32;
 import org.freedesktop.dbus.types.Variant;
 import org.freedesktop.modemmanager1.Modem;
 import org.freedesktop.modemmanager1.modem.Location;
@@ -72,7 +71,6 @@ public class NMDbusConnector {
     private static final String NM_SETTINGS_BUS_PATH = "/org/freedesktop/NetworkManager/Settings";
     private static final String MM_BUS_NAME = "org.freedesktop.ModemManager1";
     private static final String MM_MODEM_NAME = "org.freedesktop.ModemManager1.Modem";
-    private static final String MM_SIM_NAME = "org.freedesktop.ModemManager1.Sim";
     private static final String MM_LOCATION_BUS_NAME = "org.freedesktop.ModemManager1.Modem.Location";
 
     private static final String NM_DEVICE_PROPERTY_INTERFACE = "Interface";
@@ -252,8 +250,9 @@ public class NMDbusConnector {
         if (modemPath.isPresent()) {
             modemDeviceProperties = this.modemManager.getModemProperties(modemPath.get());
             if (modemDeviceProperties.isPresent()) {
-                simProperties = getModemSimProperties(modemDeviceProperties.get());
-                bearerProperties = getModemBearersProperties(modemPath.get(), modemDeviceProperties.get());
+                simProperties = this.modemManager.getModemSimProperties(modemDeviceProperties.get());
+                bearerProperties = this.modemManager.getModemBearersProperties(modemPath.get(),
+                        modemDeviceProperties.get());
             }
         }
         DevicePropertiesWrapper modemPropertiesWrapper = new DevicePropertiesWrapper(deviceProperties,
@@ -655,77 +654,4 @@ public class NMDbusConnector {
         }
     }
 
-    private List<SimProperties> getModemSimProperties(Properties modemProperties) throws DBusException {
-        List<SimProperties> simProperties = new ArrayList<>();
-        try {
-            UInt32 primarySimSlot = modemProperties.Get(MM_MODEM_NAME, "PrimarySimSlot");
-
-            if (primarySimSlot.intValue() == 0) {
-                // Multiple SIM slots aren't supported
-                DBusPath simPath = modemProperties.Get(MM_MODEM_NAME, "Sim");
-                if (!simPath.getPath().equals("/")) {
-                    Properties simProp = this.dbusConnection.getRemoteObject(MM_BUS_NAME, simPath.getPath(),
-                            Properties.class);
-                    simProperties.add(new SimProperties(simProp, true, true));
-                }
-            } else {
-                List<DBusPath> simPaths = modemProperties.Get(MM_MODEM_NAME, "SimSlots");
-                for (int index = 0; index < simPaths.size(); index++) {
-                    String dbusPath = simPaths.get(index).getPath();
-
-                    if (dbusPath.equals("/")) {
-                        // SIM slot doesn't contain a SIM
-                        continue;
-                    }
-
-                    Properties simProp = this.dbusConnection.getRemoteObject(MM_BUS_NAME, dbusPath, Properties.class);
-                    boolean isActive = simProp.Get(MM_SIM_NAME, "Active");
-                    boolean isPrimary = index == (primarySimSlot.intValue() - 1);
-
-                    simProperties.add(new SimProperties(simProp, isActive, isPrimary));
-                }
-            }
-        } catch (DBusExecutionException e) {
-            // Fallback for ModemManager version prior to 1.16
-            DBusPath simPath = modemProperties.Get(MM_MODEM_NAME, "Sim");
-            if (!simPath.getPath().equals("/")) {
-                Properties simProp = this.dbusConnection.getRemoteObject(MM_BUS_NAME, simPath.getPath(),
-                        Properties.class);
-                simProperties.add(new SimProperties(simProp, true, true));
-            }
-
-        }
-        return simProperties;
-    }
-
-    private List<Properties> getModemBearersProperties(String modemPath, Properties modemProperties)
-            throws DBusException {
-        List<Properties> bearerProperties = new ArrayList<>();
-        try {
-            Modem modem = this.dbusConnection.getRemoteObject(MM_BUS_NAME, modemPath, Modem.class);
-            if (Objects.nonNull(modem)) {
-                List<DBusPath> bearerPaths = modem.ListBearers();
-                bearerProperties = getBearersPropertiesFromPaths(bearerPaths);
-            }
-        } catch (DBusExecutionException e) {
-            try {
-                List<DBusPath> bearerPaths = modemProperties.Get(MM_BUS_NAME, "Bearers");
-                bearerProperties = getBearersPropertiesFromPaths(bearerPaths);
-            } catch (DBusExecutionException e1) {
-                logger.warn("Cannot get bearers for modem {}", modemPath, e1);
-            }
-        }
-        return bearerProperties;
-    }
-
-    private List<Properties> getBearersPropertiesFromPaths(List<DBusPath> bearerPaths) throws DBusException {
-        List<Properties> bearerProperties = new ArrayList<>();
-        for (DBusPath bearerPath : bearerPaths) {
-            if (!bearerPath.getPath().equals("/")) {
-                bearerProperties
-                        .add(this.dbusConnection.getRemoteObject(MM_BUS_NAME, bearerPath.getPath(), Properties.class));
-            }
-        }
-        return bearerProperties;
-    }
 }
